requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
langgraph
langchain-huggingface
faiss-cpu
sentence-transformers

tavily-python
python-dotenv
requests
beautifulsoup4

pypdf
unstructured
unstructured[pdf]
pandas
tqdm
fastapi
uvicorn[standard]